jcw/apps/api/requirements.txt
Django==5.0.*
djangorestframework==3.14.*
django-tenants==3.6.*
django-environ==0.11.*
psycopg[binary]==3.1.*
django-cors-headers==4.3.*
djangorestframework-simplejwt==5.3.*
Pillow==10.1.*